active record integration complete; active record migration complete; mongoid migration started; NO mongoid integration

+++ spec/spec_helper.rb
@@ -1,12 +1,21 @@
 # encoding: utf-8
+# Now there are support only for "active_record"
 ENV["ORM"] = "active_record"
-USE_RAILS = false
+# Set this to true if you want to launch rails application
+USE_RAILS = true
+# Set this true to use debugger, if your  ruby version supports debugger.
 USE_DEBUGGER = true
-require 'header'
+# Set this to true to see HTML code coverage report
+SHOW_REPORT = false
 
+
+require 'header'
 if USE_RAILS
-  require 'ar_schema'
   require 'rails_helper'
+end
+
+if ENV["ORM"] == "active_record"
+  require 'ar_schema'
 end
 
 if USE_DEBUGGER
@@ -22,6 +31,7 @@ require File.expand_path('lib/lolita-translation')
 I18n.available_locales = [:en,:lv,:ru,:fr]
 I18n.default_locale = :lv
 I18n.locale = :en
+Lolita.locales = I18n.available_locales
 
 
 RSpec.configure do |config|
@@ -40,5 +50,7 @@ at_exit do
       File.delete(File.expand_path("spec/test_app/db/lolita-translation.db"))
     end
   end
-  #CoverMe.complete!
+  if ::SHOW_REPORT
+    CoverMe.complete!
+  end
 end
